Remove abandoned module `appchain_message_processing_results`.
